requirements.txt
Django>=4.2.0
python-dotenv>=1.0.0
supabase>=2.0.0
requests>=2.31.0
twilio>=8.12.0
django-webpush>=0.3.5  # For web push notifications
pytz>=2024.1  # For timezone handling 